add base interface class for data importer plugins and began conversion of the current importer.
Created basic Title class that can be serialized so that we dont have to read file data every time we load up, should be alot faster (think one filehandle and binary data, this removes the need for data parsing on each load).

Also began the process of converting the Movie class to using Title objects instead of the dataSet concept.