Ok, MAJOR commit here.  I have now modularized the codebase.  Library (which I'd like to rename to OpenMediaLibrary at some point) ONLY contains mcml and DisplayItem type stuff, ok, and a utility method for loading images.
OMLEngine: is a new project that can be worked on independantly that contains the real guts of the code.  This project (c# library) contains the Title, titleCollection, and other main classes.  This includes the data importer interface and the current code of 2 data importers (DVDProfilerImporter and MovieCollectorzImporter).  These will at some point likely be moved to their own projects as well so that they directly compile into dll files.

OMLDatabaseEditor: is a generic windows gui app that I'll now being working on directly to manage the database.

OMLEngineTestSuite: is a general test suite package for testing all the relevant parts (OMLEngine and OMLDatabaseEditor for now).  This can be expanded at any time and should be!  Remeber: unit tests are a good thing!  Some of the tests currently fail because I haven't yet fleshed out the tests for things like serialization verification and the like.