I'm looking into adding a new Importer.  This importer would be a generic OML oriented XML file parser.  The XML file could contain one or more movie descriptions, but I think it might be more specific to a given movie.  Anyway this is a sample of what I'm thinking might be in the file.  Although much of it might be optional, I would like to set the Title, Location and Format as the bare minimum.  This should, I believe, be enough to get the movie into OML.
In the long run the files might be called [moviename].oml.xml; where [moviename] is the name of the movie.  If so, then the importer only needs the location, and like the dvr-ms importer it will look for the front cover as [moviename].jpg or location with a modified .jpg extension.