Sort should now work correctly for movies.  I just changed it to the sort method on the title collection class instead of using the trees.
titlenames in tree should now change when saved.

datagridviews now are updating without you having to change focus.

I also added a call to the raw setup in my constructor.

I was also playing around for a little bit with copy and paste but didn't get it working so I just commented it all out.