Added bug fix based on post on TGB by MostlyHarmless "Here's a bug in the DVD Library XML importer (and a fix):
My issue was the importer was only finding 383 out of 405 DVDs.

In DVDLibraryImporter.cs, function GetDVDCacheFileName, the problem was some of my DVDIDs weren't matching the expression dvdid.Length == 17 && dvdid[8 ] == '|'  So it ended up not replacing the pipe character with a dash, and then a filenotfound exception was thrown since it wouldn't find the file. DVDIDs are not always 17 characters in length. Some of my DVDIDs are only 15 or 16 characters long (only 6 or 7 characters before the pipe)."