adding the ability to click on an actor in the actor panel to see all the movies they're in
because of the way filters work with pre-filtering all the views - i needed to add code to add a filter after all the filtering has been done.  so if the user doesn't have the "actor" filter - they can still navigate there - i just create it when i notice it's not there by scanning through all the titles again.

still todo is make the directors work as well