fixing a couple bugs with Mounting images
this will break mounting over the extender -  i'll need to work with the extender guys to figure out the correct thing

basically the mounting tool was updating the Title and turning it into dvd because things are passed by REF 

i've removed the idea of a mounting player and now just mount when creating a player and send the mounted path into the various players.

this will allow mounting to work alongside the external player so you can mount a bluray and an external player play it.