updating the sql script to use INTS for the title id - this helps with bootstrapping the id as the unique id - once everything is more moved over we can talk about changing the type
adding more boostrapping/prototyping as a proof of concept.  the initial load in the UI is now grabbed from the DB.  when you go to the details view you always load from the DB and i grab the people data as well.  updating the watched/unwatched count is a direct db write now.

going to the "unwatched" filter is now grabbed as a db query every time.

right now there is a no relationship between the title/genre/etc objects and the db objects - i create new objects out of the db objects.  changing this is a big design step i'll try to tackle soon.

i haven't thought how embedded queries should work yet ie ( filtering on genre and year and length ) - maybe a collection of where clauses we could pass into a method and have a select return our records.  this gets stickier when you're in the view that tells you all the filters and their record counts.

i don't think the filtering on all people is going to be possible - it's just too slow.  with my data that view isn't usable anyways since it lists 17k people.  the filtering on an individual user from the details view should still be reasonable howerver.