updated the sql schema to: spell CountryOfOrigin correctly added OriginalName added ImporterSource
made sortname not required

The title object now contains a DB title object and uses thunking to get expose the properties..i walked down the path of using partial classes and it's a real headache - plus it exposes stuff we may not want to expose and the collection classes become tricky.  The thunking layer with containment is probably the right thing to do.

a lot of the functionality inside title is remarked out - mostly having to do with serializing the object to/from xml.  I'm not sure this functionality is needed anymore since we can just script out the DB if we need a file version for backing up or what not.