adding more filters
ParentRating
Runtime
VideoFormat

i think the way i'm doing the sql query is bad..hopefully someone can optimize it.  for runtime i'm just grabbing all the titles and iterating through instead of doing it in sql.  we should figure out the better approach.

i created a titleconfig class that contains all the runtime strings and lengths - i'm not sure the right place for this object but i'm still flushing stuff out.  it would make sense to put the DateAdded filter stuff in there as well.