making a file system scanner that's part of the OML Engine.
the idea is to move most of the logic that exists in the DVDLibraryImporter plugin to the OMLEngine so that the setup program, importer and file system scanner service can all make use of it.  all the file system related classes should be centralized here.

adding methods to tell if a path is unique or not in the database.

adding a network library i found online that lets you enumerator all the available network shares


given the above you can now scan a set of folders and get back all the valid media items and send that into the database and get back all the media items that are new.