adding a metadatasourceid to the database - this will store the id of the metadata item and allow for checking of meta data duplicates.
getting in the beginnings of a new filewatcher

You give the watcher a list of directories to watch - which can also be network shares

when the internal filewatcher triggers a change it waits 15 seconds - every further change pushes the timer out another 15 seconds.

when the timer elapses it scans ALL the folders and lists all the media it finds.

it sends that media into the SQL database and asks if any of the media paths are new

if there are new paths it sends the name of the new items into the meta data getter ( this part is TBD )

and add the titles to the collection

The plan is to have a settings page that will allow the user to pick paths to watch and their metadata source - with DVDProfiler being an option.