Ok, new build requirement. BuildVersionUpdater is a new dependency of the library.dll project. This app will determine the assembly version of the library.dll project, it will then take this value and update both wix installers as well as the mce registration file.  This will allow the auto-updaters to understand what build version you are on and if a new version is available or not. This will also allow us to have reasonable build numbers for the vmc plugin lib.
So, when you are going to release a new build (either official or daily)

you MUST!! do the following:
edit the library project and make assembly and file version changes
to increment the version number(s).
Build the entire project.

Build and upload the msi file(s).

Edit the OMLInfo.xml file with the new correct version number
for either beta or release depending on what you are releasing.

Then just upload the new omlinfo.xml file to whatever we're putting them
and your off to the races.

Users will active updaters will auto detect the new version 
the next time they run OML and it should be able to update them.

Now I just have to finish up the WiX installers so they dont require
a reboot and this feature will be done!