Added a bootstrapper to each of the installers.
This bootstrapper will create a setup.exe for each of the msi files.
The msi files will NOT be released directly.  The setup.exe file(s) will have to be renamed
and they will have to be released instead of the msi files.

The primary purpose of the bootstrapped files is that they
are able to detect and if missing download/install required components
such as windows installer v3.1, .net 3.5 and soon ms sql express 2k8