Database installation/creation wix is now complete.
This still does not install ms sql server directly but it will fail to install
if sql server is not installed and running.

Also, there is a new library setting called HasRunSetup that defaults to false

This causes oml to goto a firstrunsetup.mcml page on first run after an install.
Then it sets this flag to true and oml will continue to run as expected after that.

We now can begin building setup screens for various items.