First Run Folders As titles.
The "Scan Folders For DVDs and Videos" scanner now supports
"folders as titles".  This means that when this option is active
it will treat multiple video files in a single folder as "disks" instead of titles.
A new title will be created using the folder name as the title name and each file
will be added to this title as a disk.

The option can be located in the DBEditor under Tools -> Options