Added option to the options form in DBEditorfor-
Added a setting 'AddParentFoldersToTitleName'. When scanning folders, if true and 'Folders are Titles' option is ticked, the scanner will add the parent folder name to the title name. Works well for tv series eg
|-Red dwarf
|     |- Series 1
|     |- Series 2
|-Cheers
|     |- Series 1
|     |- Series 2

etc.

The titles will be named
Red Dwarf - Series 1
etc.

The parent is not added for root folders.