Rolled ui-changes branch into trunk.
Notes:
To view the v3 ui you must un-rem the LAYOUT_V3 symbol at the top of Library\OMLApplication.cs

This is just a wireup of the v3 UI to a sample set of records. 
Mapping to the full entity model was on hold till SQL integration.
Settings under v3 has been disabled and is waiting for mjp to handle the merge of base control changes.