moving a ton of settings into SQL
There's now a static file in OMLEngine called OMLSettings which has all the settings for the entire application it.

You'll need to update your SQL to get the new table

All the settings are simply Name/Value pairs with a third column for "InstanceName" - the instance name will allow us to have different settings for different clients.  The idea is that when you install OML we'll write to the registry a GUID - and that will be your instance name.  Any settings you update will be updated under that instance or you'll fallback to the default instance. [ this is all to be implemented later ]

there are still some UI related settings not moved into SQL settings yet because they don't make sense.  Please review your settings and see where they fit best.

i need to spend some time figuring out the caching - i'm not sure if it's doing caching for us..and if it's not.. we'll need to add our own.