Creating a new ImageMapping table that will allow us to map images to titles in different ways.  Currently the columns are : TitleId, ImageId, ImageType
we can add more columns later if we want to store where the image came from and it's original Name

ImageType can be FrontCoverImage, BackCoverImage, FanartImage.  More image types can be added if we want to associate more images with a Title.

With this change we now store fan art in the database as well.

I removed all the image columns from the Title object since the new ImageMapping table takes care of that now

when downloading fanart images i now download them to \OML\DownloadImages as a guid before storing them in the database.  the ImageManager.CleanupCachedImages will clear out this folder when called.  This has the happy side effect of being able to pulldown the fanart before you have a disk seleceted in the db editor.

the DBEditor needs to be updated to better deal with this new design :)