fixes the multiple threads accessing the database at the same time issue ms bob was seeing
some refactor on the call pattern.  i wanted to make it more re-usable so now i have a :

Title.BeginGetFrontCoverMenuPath(DelegateFrontCoverPath callback)

Which will invoke your delegate when the cover is done being loaded.  underneath it still uses the threadpool and lazy creates it only when needed.

this approach is creating a new db connection for every call we make inside a spawned thread.  there's probably a better approach where we can re-use a pool of connections instead of creating/tearing down so many.  i personally found the perf acceptable, but if we run into an issue that may be something to look into.