having the Runtime and DateAdded groups use the TitleManager calls that return the titles grouped together
although this approach is cleaner the perf ends up being about the same since we still need a SQL call per filter in the group.  It would probably we worthwhile for someone smarter than me to figure out a way to do in a single query.  At least this refactor will make that easier.