DBEditor2 -
Changes to improve the robustness of updating meta data of a title. In some cases an exception would occur and as the metadata update is wrapped in a try catch block there is no evidence of the problem apart from missing data.

Changed the ClearEditor function to accept a parameter. If parameter true the title editor would be cleared, if not only save state flag is cleared. Allows saving a title without needing to reselect it to view changes.