turning TitleType into a filter rather than a special case
this gives the UI a lot of power to how it wants to shape what it views.  this change will take some UI massaging before things work out how they expect but the API power should be there.

this checkin includes a big hack in the TitlePivot.cs that shows how something could work but needs to be cleaned up