adding in some more fixes to the directory scanner
adding types as unknown

added a video type for the previous unknown type

removed the custom settings stuff for the directory scanner and using the other settings code.

added the ability to turn of caching of settings ( this is needed because i need to continuously query to see if someone has turned off the scanner while the scanner is running.

added a test app so you can use the scanner without installing the service