Changes to improve speed of meta data lookups.
The plugins no longer download the thumbnails, this is now handled by the search results dialog in the editor. The editor now downloads the thumbnails asynchronously up to a maximum of five concurrent downloads. Massive speed improvements in some of the plugins

User can now set the number of results to return.